separate social_capital and actions
